Modified lore, added right click, added sound
